public interfaces for graph
