+++ libs/langgraph/langgraph/graph/__init__.py
@@ -2,11 +2,11 @@ from langgraph.constants import END, START
 from langgraph.graph.message import MessageGraph, MessagesState, add_messages
 from langgraph.graph.state import StateGraph
 
-__all__ = [
+__all__ = (
     "END",
     "START",
     "StateGraph",
-    "MessageGraph",
     "add_messages",
     "MessagesState",
-]
+    "MessageGraph",
+)
