+++ libs/langgraph/langgraph/graph/message.py
@@ -27,6 +27,12 @@ from typing_extensions import TypedDict
 from langgraph.constants import CONF, CONFIG_KEY_SEND
 from langgraph.graph.state import StateGraph
 
+__all__ = (
+    "add_messages",
+    "MessagesState",
+    "MessageGraph",
+)
+
 Messages = Union[list[MessageLikeRepresentation], MessageLikeRepresentation]
 
 REMOVE_ALL_MESSAGES = "__remove_all__"
