+++ libs/langgraph/langgraph/graph/state.py
@@ -14,8 +14,6 @@ from typing import (
     Callable,
     Generic,
     Literal,
-    NamedTuple,
-    Protocol,
     Union,
     cast,
     get_args,
@@ -26,7 +24,7 @@ from typing import (
 
 from langchain_core.runnables import Runnable, RunnableConfig
 from pydantic import BaseModel
-from typing_extensions import Self, TypeAlias, Unpack
+from typing_extensions import Self, Unpack
 
 from langgraph._typing import UNSET, DeprecatedKwargs
 from langgraph.cache.base import BaseCache
@@ -56,7 +54,8 @@ from langgraph.errors import (
     ParentCommand,
     create_error_message,
 )
-from langgraph.graph.branch import Branch
+from langgraph.graph._branch import BranchSpec
+from langgraph.graph._node import StateNode, StateNodeSpec
 from langgraph.managed.base import (
     ManagedValueSpec,
     is_managed_value,
@@ -76,9 +75,8 @@ from langgraph.types import (
     Command,
     RetryPolicy,
     Send,
-    StreamWriter,
 )
-from langgraph.typing import InputT, OutputT, StateT, StateT_contra
+from langgraph.typing import InputT, OutputT, StateT
 from langgraph.utils.fields import (
     get_cached_annotated_keys,
     get_field_default,
@@ -88,7 +86,11 @@ from langgraph.utils.pydantic import create_model
 from langgraph.utils.runnable import coerce_to_runnable
 from langgraph.warnings import LangGraphDeprecatedSinceV05
 
+__all__ = ("StateGraph", "CompiledStateGraph")
+
 logger = logging.getLogger(__name__)
+
+_CHANNEL_BRANCH_TO = "branch:to:{}"
 
 
 def _warn_invalid_state_schema(schema: type[Any] | Any) -> None:
@@ -103,86 +105,11 @@ def _warn_invalid_state_schema(schema: type[Any] | Any) -> None:
     )
 
 
-class _StateNode(Protocol[StateT_contra]):
-    def __call__(self, state: StateT_contra) -> Any: ...
-
-
-class _NodeWithConfig(Protocol[StateT_contra]):
-    def __call__(self, state: StateT_contra, config: RunnableConfig) -> Any: ...
-
-
-class _NodeWithWriter(Protocol[StateT_contra]):
-    def __call__(self, state: StateT_contra, *, writer: StreamWriter) -> Any: ...
-
-
-class _NodeWithStore(Protocol[StateT_contra]):
-    def __call__(self, state: StateT_contra, *, store: BaseStore) -> Any: ...
-
-
-class _NodeWithWriterStore(Protocol[StateT_contra]):
-    def __call__(
-        self, state: StateT_contra, *, writer: StreamWriter, store: BaseStore
-    ) -> Any: ...
-
-
-class _NodeWithConfigWriter(Protocol[StateT_contra]):
-    def __call__(
-        self, state: StateT_contra, *, config: RunnableConfig, writer: StreamWriter
-    ) -> Any: ...
-
-
-class _NodeWithConfigStore(Protocol[StateT_contra]):
-    def __call__(
-        self, state: StateT_contra, *, config: RunnableConfig, store: BaseStore
-    ) -> Any: ...
-
-
-class _NodeWithConfigWriterStore(Protocol[StateT_contra]):
-    def __call__(
-        self,
-        state: StateT_contra,
-        *,
-        config: RunnableConfig,
-        writer: StreamWriter,
-        store: BaseStore,
-    ) -> Any: ...
-
-
-# TODO: we probably don't want to explicitly support the config / store signatures once
-# we move to adding a context arg. Maybe what we do is we add support for kwargs with param spec
-# this is purely for typing purposes though, so can easily change in the coming weeks.
-StateNode: TypeAlias = Union[
-    _StateNode[StateT_contra],
-    _NodeWithConfig[StateT_contra],
-    _NodeWithWriter[StateT_contra],
-    _NodeWithStore[StateT_contra],
-    _NodeWithWriterStore[StateT_contra],
-    _NodeWithConfigWriter[StateT_contra],
-    _NodeWithConfigStore[StateT_contra],
-    _NodeWithConfigWriterStore[StateT_contra],
-    Runnable[StateT_contra, Any],
-]
-
-
 def _get_node_name(node: StateNode) -> str:
     try:
         return getattr(node, "__name__", node.__class__.__name__)
     except AttributeError:
         raise TypeError(f"Unsupported node type: {type(node)}")
-
-
-class StateNodeSpec(NamedTuple):
-    # TODO: rename this callable, also move away from NamedTuple so that we can use
-    # a generic StateNode, so maybe a dataclass
-    runnable: StateNode
-    metadata: dict[str, Any] | None
-    # TODO: rename to input_schema, though we really just want to modify this structure to
-    # be a dataclass
-    input: type[Any]
-    retry_policy: RetryPolicy | Sequence[RetryPolicy] | None
-    cache_policy: CachePolicy | None
-    ends: tuple[str, ...] | dict[str, str] | None = EMPTY_SEQ
-    defer: bool = False
 
 
 class StateGraph(Generic[StateT, InputT, OutputT]):
@@ -239,7 +166,7 @@ class StateGraph(Generic[StateT, InputT, OutputT]):
 
     edges: set[tuple[str, str]]
     nodes: dict[str, StateNodeSpec]
-    branches: defaultdict[str, dict[str, Branch]]
+    branches: defaultdict[str, dict[str, BranchSpec]]
     channels: dict[str, BaseChannel]
     managed: dict[str, ManagedValueSpec]
     schemas: dict[type[Any], dict[str, BaseChannel | ManagedValueSpec]]
@@ -538,7 +465,7 @@ class StateGraph(Generic[StateT, InputT, OutputT]):
         self.nodes[node] = StateNodeSpec(
             coerce_to_runnable(action, name=node, trace=False),
             metadata,
-            input=input_schema or self.state_schema,
+            input_schema=input_schema or self.state_schema,
             retry_policy=retry_policy,
             cache_policy=cache_policy,
             ends=ends,
@@ -641,7 +568,7 @@ class StateGraph(Generic[StateT, InputT, OutputT]):
                 f"Branch with name `{path.name}` already exists for node `{source}`"
             )
         # save it
-        self.branches[source][name] = Branch.from_path(path, path_map, True)
+        self.branches[source][name] = BranchSpec.from_path(path, path_map, True)
         if schema := self.branches[source][name].input_schema:
             self._add_schema(schema)
         return self
@@ -994,7 +921,7 @@ class CompiledStateGraph(
                 writers=[ChannelWrite(write_entries)],
             )
         elif node is not None:
-            input_schema = node.input if node else self.builder._state_schema
+            input_schema = node.input_schema if node else self.builder._state_schema
             input_channels = list(self.builder.schemas[input_schema])
             is_single_input = len(input_channels) == 1 and "__root__" in input_channels
             if input_schema in self.schema_to_mapper:
@@ -1003,7 +930,7 @@ class CompiledStateGraph(
                 mapper = _pick_mapper(input_channels, input_schema)
                 self.schema_to_mapper[input_schema] = mapper
 
-            branch_channel = CHANNEL_BRANCH_TO.format(key)
+            branch_channel = _CHANNEL_BRANCH_TO.format(key)
             self.channels[branch_channel] = (
                 LastValueAfterFinish(Any)
                 if node.defer
@@ -1031,7 +958,7 @@ class CompiledStateGraph(
             if end != END:
                 self.nodes[starts].writers.append(
                     ChannelWrite(
-                        (ChannelWriteEntry(CHANNEL_BRANCH_TO.format(end), None),)
+                        (ChannelWriteEntry(_CHANNEL_BRANCH_TO.format(end), None),)
                     )
                 )
         elif end != END:
@@ -1052,7 +979,7 @@ class CompiledStateGraph(
                 )
 
     def attach_branch(
-        self, start: str, name: str, branch: Branch, *, with_reader: bool = True
+        self, start: str, name: str, branch: BranchSpec, *, with_reader: bool = True
     ) -> None:
         def get_writes(
             packets: Sequence[str | Send], static: bool = False
@@ -1060,7 +987,7 @@ class CompiledStateGraph(
             writes = [
                 (
                     ChannelWriteEntry(
-                        p if p == END else CHANNEL_BRANCH_TO.format(p), None
+                        p if p == END else _CHANNEL_BRANCH_TO.format(p), None
                     )
                     if not isinstance(p, Send)
                     else p
@@ -1075,7 +1002,7 @@ class CompiledStateGraph(
         if with_reader:
             # get schema
             schema = branch.input_schema or (
-                self.builder.nodes[start].input
+                self.builder.nodes[start].input_schema
                 if start in self.builder.nodes
                 else self.builder.state_schema
             )
@@ -1237,12 +1164,12 @@ def _control_branch(value: Any) -> Sequence[tuple[str, Any]]:
         if isinstance(command.goto, Send):
             rtn.append((TASKS, command.goto))
         elif isinstance(command.goto, str):
-            rtn.append((CHANNEL_BRANCH_TO.format(command.goto), None))
+            rtn.append((_CHANNEL_BRANCH_TO.format(command.goto), None))
         else:
             rtn.extend(
                 (TASKS, go)
                 if isinstance(go, Send)
-                else (CHANNEL_BRANCH_TO.format(go), None)
+                else (_CHANNEL_BRANCH_TO.format(go), None)
                 for go in command.goto
             )
     return rtn
@@ -1253,12 +1180,12 @@ def _control_static(
 ) -> Sequence[tuple[str, Any, str | None]]:
     if isinstance(ends, dict):
         return [
-            (k if k == END else CHANNEL_BRANCH_TO.format(k), None, label)
+            (k if k == END else _CHANNEL_BRANCH_TO.format(k), None, label)
             for k, label in ends.items()
         ]
     else:
         return [
-            (e if e == END else CHANNEL_BRANCH_TO.format(e), None, None) for e in ends
+            (e if e == END else _CHANNEL_BRANCH_TO.format(e), None, None) for e in ends
         ]
 
 
@@ -1415,6 +1342,3 @@ def _get_schema(
                     if k in channels and isinstance(channels[k], BaseChannel)
                 },
             )
-
-
-CHANNEL_BRANCH_TO = "branch:to:{}"
